% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Structstrings-LoopIndexList.R
\docType{class}
\name{LoopIndexList}
\alias{LoopIndexList}
\alias{LoopIndexList-class}
\title{LoopIndexList: base pairing information as a list of integer values}
\usage{
LoopIndexList(...)
}
\arguments{
\item{...}{the \code{integer} input vectors.}
}
\value{
a \code{LoopIndexList} object.
}
\description{
With loop indeces base pairing information can be represented by giving each base
pair a number and increasing/decreasing it with each opened/closed base
pair. This information can be used for further analysis of the represented 
structure.
}
\examples{
# if the object is create manually make sure it is a valid structure
# information. Otherwise an error is thrown.
lil <- LoopIndexList(list(c(1L,2L,3L,3L,3L,2L,1L,0L,5L,6L,6L,5L),
                       c(1L,2L,2L,2L,2L,2L,1L,0L,5L,6L,6L,5L)))
}
