% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Structstrings.R
\docType{package}
\name{Structstrings-package}
\alias{Structstrings-package}
\title{Structstrings: Implementation of the dot bracket annotations with Biostrings}
\description{
The Structstrings package implements the widely used dot bracket annotation for storing base pairing information in structured RNA. Structstrings uses the infrastructure provided by the Biostrings package and derives the DotBracketString and related classes from the BString class. From these, base pair tables can be produced for in depth analysis. In addition, the loop indices of the base pairs can be retrieved as well. For better efficiency, information conversion is implemented in C, inspired to a large extend by the ViennaRNA package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/FelixErnst/Structstrings}
  \item Report bugs at \url{https://github.com/FelixErnst/Structstrings/issues}
}

}
\author{
\strong{Maintainer}: Felix G.M. Ernst \email{felix.gm.ernst@outlook.com} (\href{https://orcid.org/0000-0001-5064-0928}{ORCID})

}
\keyword{internal}
