% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Structstrings.R
\docType{data}
\name{Structstrings-data}
\alias{Structstrings-data}
\alias{dbs}
\alias{seq}
\title{Structstrings example data}
\format{object of class \code{\link{DotBracketStringSet}}}
\usage{
Structstrings

seq
}
\description{
Example data for using the Structstrings package
}
\keyword{Structstrings}
\keyword{datasets}
