% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Structstrings.R
\docType{data}
\name{Structstrings-data}
\alias{Structstrings-data}
\alias{dbs}
\alias{nseq}
\title{Structstrings example data}
\format{
object of class \code{\link{DotBracketStringSet}} and 
\code{\link{DNAStringSet}}

An object of class \code{DNAStringSet} of length 299.
}
\source{
sequence and  dot bracket annotation of tRNAscan-SE output for 
*S. cerevisiae* imported using 
\code{\link[tRNAscanImport:tRNAscanImport]{tRNAscanImport}}. The example file
is part of the \code{tRNAscanImport} package.
}
\usage{
data(dbs)

data(nseq)
}
\description{
Example data for using the Structstrings package
}
\keyword{datasets}
