% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Structstrings.R,
%   R/Structstrings-DotBracketDataFrame.R
\docType{data}
\name{Structstrings-internals}
\alias{Structstrings-internals}
\alias{DOTBRACKET_CHAR_VALUES}
\alias{DOTBRACKET_ALPHABET}
\alias{STRUCTURE_NEUTRAL_CHR}
\alias{STRUCTURE_OPEN_CHR}
\alias{STRUCTURE_CLOSE_CHR}
\alias{[<-,DotBracketDataFrame-method}
\alias{colnames<-,CompressedSplitDotBracketDataFrameList-method}
\title{Structstrings internals}
\format{An object of class \code{integer} of length 9.}
\usage{
DOTBRACKET_CHAR_VALUES

DOTBRACKET_ALPHABET

STRUCTURE_NEUTRAL_CHR

STRUCTURE_OPEN_CHR

STRUCTURE_CLOSE_CHR

\S4method{[}{DotBracketDataFrame}(x, i, j, ...) <- value

\S4method{colnames}{CompressedSplitDotBracketDataFrameList}(x) <- value
}
\arguments{
\item{x, i, j, ..., value}{See \link{DataFrame}.}
}
\description{
Analog to \code{Biostrings} there are a few objects, which should only be 
used internally, but may be of use to other package developers.
Otherwise take care.
}
\examples{
DOTBRACKET_CHAR_VALUES
DOTBRACKET_ALPHABET
STRUCTURE_NEUTRAL_CHR
STRUCTURE_OPEN_CHR
STRUCTURE_CLOSE_CHR

# the replace method for a DotBracketDataFrame had to be reimplemented
# because of the requirement of columns for a DotBracketDataFrameList and
# DotBracketDataFrame
data("dbs", package = "Structstrings", envir = environment())
dbdfl <- getBasePairing(dbs)
# Elements are returned as DotBracketDataFrames
dbdf <- dbdfl[[1]]
dbdfl[[1]] <- dbdf
dbdfl[1] <- dbdfl[1]
}
\keyword{datasets}
